requirements.txt
# System deps (Ubuntu/Debian):
# sudo apt install -y portaudio19-dev libsndfile1 python3-dev build-essential

faiss-cpu==1.13.2
numpy==2.2.6
openai==2.14.0
opencv-python==4.12.0.88
sounddevice==0.4.7
langdetect==1.0.9
websocket-client==1.8.0
